backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncio==3.4.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
docstring-parser==0.16
executing==2.2.0
fastapi==0.115.12
git-filter-repo==2.47.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.9.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.2
pydantic-core==2.33.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyzmq==26.4.0
requests==2.32.3
rsa==4.9
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
